Reimplemented the waitForAsyncResult function to now wait on all AsyncResult values in nested builtin structures (list, tuple, dict).
This means that if the argument to a function decorated with @waitForAsyncResult is a list of AsyncResult values then it will now wait until those nested AsyncResult values are available before calling the function.

Greatly simplified the logic of flatten() as a result.

Added some unit-tests for waitForAsyncResult() and flatten().