The set of tools are now defined in Script.tools rather than using a threading.local() object as the 'cake.tools' module. A proxy object has been setup as 'cake.tools' that looks up the tools in the Script.getCurrent().tools dictionary.
This fixes the problem where sometimes to perform 'from cake.tools import mytool' in a callback that may be executed after the script finishes executing caused it to fail to find the tools.
Since the Script.getCurrent() context is propagated to child tasks of the script, the cake.tools context should now also be propagated to child tasks.