Fixes issue #17 - Using ProjectTool fails on non-Windows platform.
The problem is the ProjectTool attempts to determine if the compiler is
MSVC, to do this it tries to import the module with the MsvcCompiler.
On platforms without the Windows Registry (like Linux), the user is faced
with an import error for _winreg.

The import error is now caught and a type is defined to satisfy the check.

At this time MsvcCompiler isn't supported outside of Microsoft Windows.